% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_table.R
\name{dth_table}
\alias{dth_table}
\title{Discrete Time Hazard Table}
\usage{
dth_table(data, exp_prefix, event_dates, dlo_max = "none")
}
\arguments{
\item{data}{input cohort (wide format)}

\item{exp_prefix}{exposure variable prefix}

\item{event_dates}{year of event variable name (s)}

\item{dlo_max}{year to extend cohort. Defaults to "none"}
}
\description{
Formats cohort data into discrete time hazard data structures, to a specified time-point.
}
\details{
the data data.frame must contain the following variables:

- id,

- start_fu (start of follow-up, date)

- dlo (date of last observation, date)

- age_sfu (age at start of follow-up, numeric)

- covariate_constant (any name, any format)
}
\examples{
dth_table()
dth_table(table_example, exp_prefix = 'acm_hours', event_dates = c("meso"), dlo_max = 2022)
}
\keyword{table}
