% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_risk.R
\name{expected_risks}
\alias{expected_risks}
\title{Expected Risk of an Event of Interest}
\usage{
expected_risks(data, hazards, id_in = "id", prefix = "")
}
\arguments{
\item{data}{data.frame of predicted value of the estimated probability of the outcome for each person-period in the discrete time hazard table}

\item{hazards}{list of column names of the predicted hazard outcomes}

\item{id_in}{String column name of the identification variable. Defaults to "id"}

\item{prefix}{String of the predicted hazard outcomes column name prefix. Defaults to ""}
}
\value{
expected events: a data.frame of rt_x ... the number of predicted events corresponding to the specified hazards,
'Estimated Person Years' total number of person years in the input data

all data: a data.frame of cumulative probability of the outcome for each person-period in the discrete time hazard table
}
\description{
Calculates the probability of surviving for each person i and time interval u.
}
\examples{
expected_risks()
expected_risks(hazards_example, hazards = c("h1", "h2","h3","h4","h5","h6","h7"),
prefix = "h", id_in = "id")
}
