% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPOSH.R
\docType{package}
\name{GPOSH}
\alias{GPOSH-package}
\alias{GPOSH}
\title{GPOSH Package}
\description{
G-computation for Policy evaluation in Occupational Safety and Health
}
\author{
Louis Everest \email{louis.everest@ontariohealth.ca}
}
